Question 클래스를 DB 테이블에 매핑
Question 클래스를 @Entity로 설정하고 QuestionRepository를 생성한다.
질문을 생성할 때마다 id 값을 부여하고 기본키로 설정한다.
