+++ src/main/java/com/codessquad/domain/QuestionRepository.java
@@ -1,0 +1,6 @@
+package com.codessquad.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface QuestionRepository extends JpaRepository<Question,Long> {
+}
