+++ src/main/java/com/codessquad/qna/QuestionController.java
@@ -1,5 +1,6 @@
 package com.codessquad.qna;
 
+import com.codessquad.domain.Question;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
